+++ R/bvhar-minnesota.R
@@ -4,7 +4,7 @@
 #' This function fits BVHAR with Minnesota prior.
 #' 
 #' @param y Time series data of which columns indicate the variables
-#' @param type Prior mean type to apply. ("VAR" = Original Minnesota. "VHAR" = fills zero in the first block)
+#' @param mn_type Prior mean type to apply. ("VAR" = Original Minnesota. "VHAR" = fills zero in the first block)
 #' @param sigma Standard error vector for each variable
 #' @param lambda Tightness of the prior around a random walk or white noise
 #' @param delta Prior belief about white noise (Litterman sets 1: default)
@@ -12,6 +12,7 @@
 #' @param weekly Fill the second part in the first block
 #' @param monthly Fill the third part in the first block
 #' @param eps Very small number
+#' @param type `r lifecycle::badge("experimental")` add constant term (`"const"`) or not (`"none"`)
 #' 
 #' @details 
 #' Apply Minnesota prior to Vector HAR: \eqn{\Phi} (VHAR matrices) and \eqn{\Sigma_e} (residual covariance).
@@ -21,13 +22,13 @@
 #' (MN: \href{https://en.wikipedia.org/wiki/Matrix_normal_distribution}{matrix normal}, IW: \href{https://en.wikipedia.org/wiki/Inverse-Wishart_distribution}{inverse-wishart})
 #' 
 #' Two types of Minnesota priors builds different dummy variables for Y0.
-#' \code{type = "VAR"} constructs dummy Y0 with \code{p = 3} of \code{\link{build_ydummy}}.
+#' `mn_type = "VAR"` constructs dummy Y0 with `p = 3` of [build_ydummy()].
 #' The only difference from BVAR is dimension.
 #' 
-#' On the other hand, dummy Y0 for \code{type = "VHAR"} has its own function \code{\link{build_ydummy_bvhar}}.
+#' On the other hand, dummy Y0 for `mn_type = "VHAR"` has its own function [build_ydummy_bvhar()].
 #' It fills the zero matrix in the first block in Bańbura et al. (2010).
 #' 
-#' @return \code{bvhar_minnesota} returns an object \code{bvharmn} \link{class}.
+#' @return [bvhar_minnesota()] returns an object `bvharmn` [class].
 #' 
 #' It is a list with the following components:
 #' 
@@ -63,40 +64,40 @@
 #' @order 1
 #' @export
 bvhar_minnesota <- function(y, 
-                            type = c("VAR", "VHAR"), 
+                            mn_type = c("VAR", "VHAR"), 
                             sigma, 
                             lambda, 
                             delta, 
                             daily, 
                             weekly, 
                             monthly, 
-                            eps = 1e-04) {
+                            eps = 1e-04, 
+                            type = c("const", "none")) {
   if (!is.matrix(y)) y <- as.matrix(y)
-  type <- match.arg(type)
+  mn_type <- match.arg(mn_type)
+  m <- ncol(y)
+  N <- nrow(y)
+  num_coef <- 3 * m + 1
   # Y0 = X0 B + Z---------------------
   Y0 <- build_y0(y, 22, 23)
   name_var <- colnames(y)
   colnames(Y0) <- name_var
   X0 <- build_design(y, 22)
-  HARtrans <- scale_har(ncol(y))
-  X1 <- X0 %*% t(HARtrans)
+  HARtrans <- scale_har(m)
   name_har <- concatenate_colnames(name_var, c("day", "week", "month")) # in misc-r.R file
-  colnames(X1) <- name_har
-  m <- ncol(y)
-  N <- nrow(y)
   # dummy-----------------------------
   Yh <- switch(
-    type,
+    mn_type,
     "VAR" = {
-      if (missing(delta)) delta <- rep(1, ncol(y))
+      if (missing(delta)) delta <- rep(1, m)
       Yh <- build_ydummy(3, sigma, lambda, delta)
       colnames(Yh) <- name_var
       Yh
     },
     "VHAR" = {
-      if (missing(daily)) daily <- rep(1, ncol(y))
-      if (missing(weekly)) weekly <- rep(1, ncol(y))
-      if (missing(monthly)) monthly <- rep(1, ncol(y))
+      if (missing(daily)) daily <- rep(1, m)
+      if (missing(weekly)) weekly <- rep(1, m)
+      if (missing(monthly)) monthly <- rep(1, m)
       Yh <- build_ydummy_bvhar(sigma, lambda, daily, weekly, monthly)
       colnames(Yh) <- name_var
       Yh
@@ -104,6 +105,19 @@ bvhar_minnesota <- function(y,
   )
   Xh <- build_xdummy(3, lambda, sigma, eps)
   colnames(Xh) <- name_har
+  # const or none---------------------
+  type <- match.arg(type)
+  if (type == "none") {
+    X0 <- X0[, -(22 * m + 1)] # exclude 1 column
+    HARtrans <- HARtrans[-num_coef, -(22 * m + 1)] # HARtrans: 3m x 22m matrix
+    Th <- nrow(Yh)
+    Yh <- Yh[-Th,] # exclude intercept block from Yh (last row)
+    Xh <- Xh[-Th, -num_coef] # exclude intercept block from Xh (last row and last column)
+    name_har <- name_har[-num_coef] # remove const (row)name
+    num_coef <- num_coef - 1 # df = 3 * m
+  }
+  X1 <- X0 %*% t(HARtrans)
+  colnames(X1) <- name_har
   # estimate-bvar.cpp-----------------
   posterior <- estimate_bvar_mn(X1, Y0, Xh, Yh)
   # Prior-----------------------------
@@ -132,7 +146,8 @@ bvhar_minnesota <- function(y,
     m = m, # m
     obs = nrow(Y0), # s = n - p
     totobs = N, # n
-    process = ifelse(type == "VAR", "BVHAR_mn_var", "BVHAR_mn_vhar"),
+    process = ifelse(mn_type == "VAR", "BVHAR_mn_var", "BVHAR_mn_vhar"),
+    type = type,
     call = match.call(),
     # HAR------------------
     HARtrans = HARtrans,