+++ R/print-bvharmn.R
@@ -9,13 +9,19 @@ print.bvharmn <- function(x, digits = max(3L, getOption("digits") - 3L), ...) {
     "Call:\n",
     paste(deparse(x$call), sep="\n", collapse = "\n"), "\n\n", sep = ""
   )
-  # split the matrix for the print: B1, ..., Bp
-  phihat_mat <- 
-    split.data.frame(x$mn_mean[-(3 * x$m + 1),], gl(3, x$m)) %>% 
-    lapply(t)
+  # split the matrix for the print: Phi(d), Phi(w), Phi(m)
+  phihat_mat <- switch(
+    x$type,
+    "const" = {
+      split.data.frame(x$mn_mean[-(3 * x$m + 1),], gl(3, x$m)) %>% 
+        lapply(t)
+    },
+    "none" = {
+      split.data.frame(x$mn_mean, gl(3, x$m)) %>% 
+        lapply(t)
+    }
+  )
   names(phihat_mat) <- c("day", "week", "month")
-  # const term
-  intercept <- x$mn_mean[3 * x$m + 1,]
   cat("BVHAR with Minnesota Prior\n")
   cat("====================================================\n\n")
   cat("Phi ~ Matrix Normal (Mean, Scale 1, Scale 2 = Sigma)\n")
@@ -31,16 +37,20 @@ print.bvharmn <- function(x, digits = max(3L, getOption("digits") - 3L), ...) {
     )
     cat("\n\n")
   }
-  cat("Matrix Normal Mean for constant part:\n")
-  # c-------------------------------
-  print.default(
-    intercept,
-    digits = digits,
-    print.gap = 2L,
-    quote = FALSE
-  )
+  # const term----------------------
+  if (x$type == "const") {
+    intercept <- x$mn_mean[3 * x$m + 1,]
+    cat("Matrix Normal Mean for constant part:\n")
+    print.default(
+      intercept,
+      digits = digits,
+      print.gap = 2L,
+      quote = FALSE
+    )
+    cat("\n\n")
+  }
   # scale matrix-------------------
-  cat("\n\ndim(Matrix Normal precision matrix):\n")
+  cat("dim(Matrix Normal precision matrix):\n")
   print.default(
     dim(x$mn_prec),
     digits = digits,