+++ src/forecast-bvhar.cpp
@@ -20,17 +20,18 @@ Rcpp::List forecast_bvharmn(Rcpp::List object, int step) {
   Eigen::MatrixXd transformed_prec_mat = HARtrans.adjoint() * posterior_prec_mat.inverse() * HARtrans; // to compute SE: play a role V in BVAR
   int dim = object["m"]; // dimension of time series
   int num_design = object["obs"]; // s = n - p
-  Eigen::MatrixXd last_pvec(1, 22 * dim + 1); // vectorize the last 22 observation and include 1
-  Eigen::MatrixXd tmp_vec(1, 22 * dim);
+  int dim_har = HARtrans.cols(); // 22m + 1 (const) or 22m (none)
+  Eigen::MatrixXd last_pvec(1, dim_har); // vectorize the last 22 observation and include 1
+  Eigen::MatrixXd tmp_vec(1, 21 * dim);
   Eigen::MatrixXd res(step, dim); // h x m matrix
   Eigen::VectorXd sig_closed(step); // se^2 for each forecast (except Sigma2 part, i.e. closed form)
   for (int i = 0; i < step; i++) {
     sig_closed(i) = 1.0;
   }
+  last_pvec(0, dim_har - 1) = 1.0;
   for (int i = 0; i < 22; i++) {
     last_pvec.block(0, i * dim, 1, dim) = response_mat.block(num_design - 1 - i, 0, 1, dim);
   }
-  last_pvec(0, 22 * dim) = 1.0;
   sig_closed.block(0, 0, 1, 1) += last_pvec * transformed_prec_mat * last_pvec.adjoint();
   res.block(0, 0, 1, dim) = last_pvec * HARtrans.adjoint() * posterior_mean_mat; // y(n + 1)^T = [y(n)^T, ..., y(n - p + 1)^T, 1] %*% t(HARtrans) %*% Phihat
   if (step == 1) {