+++ NEWS.md
@@ -1,5 +1,7 @@
 # bvhar 0.2.0
 
+* Added constant term choice in each function (`var_lm()`, `vhar_lm()`, `bvar_minnesota()`, `bvar_flat()`, and `bvhar_minnesota()`).
+
 # bvhar 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.